outras melhorias
