+++ the-globoplay/the-globoplay/Scenes/SearchMovieViewController.swift
@@ -31,7 +31,5 @@ class SearchMovieViewController: UIViewController {
 
         }
         
-        
     }
-    
 }
