+++ the-globoplay/the-globoplay/Scenes/ViewController.swift
@@ -7,9 +7,7 @@
 
 import UIKit
 
-
 class ViewController: UIViewController {
-    
     
     private var request = Service()
     var genres = [] as [Genre]
